+++ src/lib/net/AddressMapper.kt
@@ -1,4 +1,4 @@
-package lib
+package lib.net
 
 import gui.GUI
 import lib.db.IPersistanceAdaptor
@@ -11,11 +11,11 @@ import java.util.*
  * Created by bensoer on 28/02/16.
  */
 
-class AddressMapper(private val dataStore:IPersistanceAdaptor){
+class AddressMapper(private val dataStore: IPersistanceAdaptor){
 
     //private var portMappings:Set<AddressPair>;
-    private var socketMappings:HashMap<SelectionKey,SocketChannel> = HashMap();
-    private var inverseSocketMappings:HashMap<SocketChannel,SelectionKey> = HashMap();
+    private var socketMappings: HashMap<SelectionKey, SocketChannel> = HashMap();
+    private var inverseSocketMappings: HashMap<SocketChannel, SelectionKey> = HashMap();
 
     init {
         //this.portMappings = dataStore.loadAll()
@@ -29,7 +29,7 @@ class AddressMapper(private val dataStore:IPersistanceAdaptor){
         this.dataStore.saveAddress(addressPair);
     }
 
-    fun deletePortMapping(addressPair:AddressPair){
+    fun deletePortMapping(addressPair: AddressPair){
         this.dataStore.deleteAddress(addressPair);
     }
 
@@ -59,7 +59,7 @@ class AddressMapper(private val dataStore:IPersistanceAdaptor){
         this.inverseSocketMappings.put(channel,key);
     }
 
-    fun getSocketChannel(key: SelectionKey):SocketChannel?{
+    fun getSocketChannel(key: SelectionKey): SocketChannel?{
         return this.socketMappings[key];
     }
